feature(examples): add all live examples

+++ libs/api/utils/src/lib/controllers/create-controllers.ts
@@ -18,9 +18,10 @@ export function createControllers<T extends mongoose.Document>(
 ) {
   return {
     // Get All
-    getAll: async () => {
+    getAll: async (userId?: string) => {
       const resources: T[] = await model
-        .find({})
+        .find()
+        .where(userId ? { userId } : {})
         .lean<T>()
         .exec();
 
@@ -28,9 +29,10 @@ export function createControllers<T extends mongoose.Document>(
     },
 
     // Get an individual resource
-    getOne: async (id: ObjectId) => {
+    getOne: async (id: ObjectId, userId?: string) => {
       const resource = await model
         .findById(id)
+        .where(userId ? { userId } : {})
         .lean<T | null>()
         .exec();
 
@@ -48,9 +50,10 @@ export function createControllers<T extends mongoose.Document>(
     },
 
     // Update a resource
-    updateOne: async (id: ObjectId, values: any) => {
+    updateOne: async (id: ObjectId, values: any, userId?: string) => {
       const resource = await model
         .findByIdAndUpdate(id, values, { new: true })
+        .where(userId ? { userId } : {})
         .lean()
         .exec();
       if (!resource)
@@ -61,9 +64,10 @@ export function createControllers<T extends mongoose.Document>(
     },
 
     // Remove one
-    removeOne: async (id: ObjectId) => {
+    removeOne: async (id: ObjectId, userId?: string) => {
       const resource = await model
         .findByIdAndRemove(id)
+        .where(userId ? { userId } : {})
         .lean()
         .exec();
